# docs/openapi.yml
openapi: 3.0.3
info:
  title: Meeting Gateway
  version: 1.0.0
  description: 'Trusted backend API. Exchange client credentials at /api/v1/auth/token/
    for a Gateway Bearer JWT over HTTPS. Gateway token default: 60 minutes (3600 seconds);
    renew using client credentials. Convay token lifetime is provider-controlled,
    observed around 6 hours; use expiresAt. Both tokens are server-to-server only.
    Scopes and ownership enforced.'
paths:
...
  /api/v1/meetings/:
    get:
      operationId: list_meetings
      description: Search only the caller’s meetings. Requires meeting:read. No bearer
        tokens or sensitive URLs returned.
      parameters:
      - in: query
        name: batch_id
        schema:
          type: string
          minLength: 1
      - in: query
        name: class_date
        schema:
          type: string
          format: date
      - in: query
        name: created_from
        schema:
          type: string
          format: date-time
      - in: query
        name: created_to
        schema:
          type: string
          format: date-time
      - in: query
        name: external_class_id
        schema:
          type: string
          minLength: 1
      - in: query
        name: id
        schema:
          type: string
          format: uuid
      - in: query
        name: page
        schema:
          type: integer
          minimum: 1
          default: 1
      - in: query
        name: page_size
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
      - in: query
        name: provider_calendar_id
        schema:
          type: string
          minLength: 1
      - in: query
        name: start_from
        schema:
          type: string
          format: date-time
      - in: query
        name: start_to
        schema:
          type: string
          format: date-time
      - in: query
        name: status
        schema:
          enum:
          - DRAFT
          - RESERVED
          - PROVISIONING
          - READY
          - LIVE
          - ENDED
          - CANCELLED
          - FAILED
          - PROVIDER_RESPONSE_INVALID
          - PROVIDER_STATE_UNKNOWN
          type: string
          minLength: 1
        description: |-
          * `DRAFT` - DRAFT
          * `RESERVED` - RESERVED
          * `PROVISIONING` - PROVISIONING
          * `READY` - READY
          * `LIVE` - LIVE
          * `ENDED` - ENDED
          * `CANCELLED` - CANCELLED
          * `FAILED` - FAILED
          * `PROVIDER_RESPONSE_INVALID` - PROVIDER_RESPONSE_INVALID
          * `PROVIDER_STATE_UNKNOWN` - PROVIDER_STATE_UNKNOWN
      - in: query
        name: teacher_id
        schema:
          type: string
          minLength: 1
      - in: query
        name: teacher_name
        schema:
          type: string
          minLength: 1
      tags:
      - api
      security:
      - GatewayBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: INVALID_TOKEN or TOKEN_EXPIRED.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: INSUFFICIENT_SCOPE.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Meeting not found for this client.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: State, room slot, capacity, or creation-in-progress conflict.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Provider contract or validation rejection.
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Provider authentication, authorization, endpoint, response,
            or unknown-state failure.
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Provider unavailable or rate limited.
    post:
      operationId: create_meeting
      description: 'Requires meeting:write. Room/license allocation is handled automatically
        by the Meeting Gateway. Obtain a Gateway Bearer token, check availability,
        choose a time, then POST metadata and startAt/endAt here. Availability is
        only a snapshot: Gateway atomically rechecks and reserves the capacity, authenticates
        internally and creates Convay before returning READY (201). Meeting creation
        is logically idempotent by IntegrationClient + class.id. If an identical scheduler
        retry is made after a safely retryable incomplete attempt, the Gateway resumes
        the same logical Meeting rather than creating a duplicate. Existing READY
        meetings are returned without another Convay creation. READY/LIVE replay and
        successful retry return 200. RETRYABLE_FAILED resumes the same record and
        reservation. Original meetingTitle, teacher.id/name, batch.id/name, class.id/date
        and startAt/endAt are immutable; changes return 409 EXISTING_MEETING_MISMATCH.
        Expired Gateway JWT returns 401 TOKEN_EXPIRED before any meeting mutation.
        RESERVED/PROVISIONING returns 409 CREATION_IN_PROGRESS; unresolved provider
        outcomes return 409 PROVIDER_RECONCILIATION_REQUIRED. Exhausted capacity returns
        409 NO_CAPACITY_AVAILABLE; retry the identical request when capacity is available.
        Timestamps must be timezone-aware and are stored in UTC. Authorization and
        startMeetingUrl require meeting:token. Use convay-token later for current
        authorization. ENDED returns 409 CLASS_ALREADY_COMPLETED; NON_RETRYABLE_FAILED
        or legacy FAILED returns 409 PREVIOUS_CREATION_FAILED; CANCELLED returns 409
        CLASS_CANCELLED. Unsupported legacy states or duplicate logical rows return
        409 EXISTING_MEETING_REQUIRES_REVIEW. These states never trigger provider
        creation.'
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMeeting'
            examples:
              MeetingCreation:
                value:
                  meetingTitle: API Test Physics
                  teacher:
                    id: T-001
                    name: Test Teacher
                  batch:
                    id: B-001
                    name: Test Batch
                  class:
                    id: CLS-001
                    date: '2026-09-24'
                  startAt: '2026-09-24T17:00:00+06:00'
                  endAt: '2026-09-24T18:00:00+06:00'
                summary: Meeting creation
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateMeeting'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateMeeting'
        required: true
      security:
      - GatewayBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponse'
              examples:
                SuccessfulMeetingCreation:
                  value:
                    success: true
                    data:
                      id: 11111111-1111-4111-8111-111111111111
                      classInfo:
                        classId: CLS-001
                        teacher:
                          id: T-001
                          name: Test Teacher
                        batch:
                          id: B-001
                          name: Test Batch
                      meetingInfo:
                        meetingTitle: API Test Physics
                        classDate: '2026-09-24'
                        scheduleType: scheduled
                        startAt: '2026-09-24T17:00:00+06:00'
                        endAt: '2026-09-24T18:00:00+06:00'
                        status: READY
                      convay:
                        meetingType: instant
                        calendarId: fake-calendar-id
                        meetingPanelAddress: meet.convay.com
                        startMeetingUrl: https://meet.convay.com/example?jwt=FAKE
                        authorization:
                          accessToken: FAKE_CONVAY_ACCESS_TOKEN
                          tokenType: Bearer
                          expiresAt: null
                      createdAt: '2026-09-24T10:00:00+06:00'
                  summary: Successful meeting creation
          description: Existing READY/LIVE meeting returned or safe scheduler retry
            completed.
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponse'
              examples:
                SuccessfulMeetingCreation:
                  value:
                    success: true
                    data:
                      id: 11111111-1111-4111-8111-111111111111
                      classInfo:
                        classId: CLS-001
                        teacher:
                          id: T-001
                          name: Test Teacher
                        batch:
                          id: B-001
                          name: Test Batch
                      meetingInfo:
                        meetingTitle: API Test Physics
                        classDate: '2026-09-24'
                        scheduleType: scheduled
                        startAt: '2026-09-24T17:00:00+06:00'
                        endAt: '2026-09-24T18:00:00+06:00'
                        status: READY
                      convay:
                        meetingType: instant
                        calendarId: fake-calendar-id
                        meetingPanelAddress: meet.convay.com
                        startMeetingUrl: https://meet.convay.com/example?jwt=FAKE
                        authorization:
                          accessToken: FAKE_CONVAY_ACCESS_TOKEN
                          tokenType: Bearer
                          expiresAt: null
                      createdAt: '2026-09-24T10:00:00+06:00'
                  summary: Successful meeting creation
          description: New meeting successfully created.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: INVALID_TOKEN or TOKEN_EXPIRED.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: INSUFFICIENT_SCOPE.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Meeting not found for this client.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                NoCapacity:
                  value:
                    success: false
                    code: NO_CAPACITY_AVAILABLE
                    message: No meeting capacity is available for the requested time.
                    details: {}
                    requestId: 22222222-2222-4222-8222-222222222222
                  summary: No capacity
          description: State, room slot, capacity, or creation-in-progress conflict.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                ProviderValidationRejection:
                  value:
                    success: false
                    code: PROVIDER_VALIDATION_ERROR
                    message: Convay rejected the meeting payload.
                    details: {}
                    requestId: 22222222-2222-4222-8222-222222222222
                  summary: Provider validation rejection
          description: Provider contract or validation rejection.
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                ProviderCreationOutcomeUnknown:
                  value:
                    success: false
                    code: PROVIDER_STATE_UNKNOWN
                    message: Provider creation outcome is unknown; reservation retained.
                    details: {}
                    requestId: 22222222-2222-4222-8222-222222222222
                  summary: Provider creation outcome unknown
          description: Provider authentication, authorization, endpoint, response,
            or unknown-state failure.
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Provider unavailable or rate limited.
  /api/v1/meetings/{id}/:
    get:
      operationId: retrieve_meeting
      description: Requires meeting:read.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - api
      security:
      - GatewayBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: INVALID_TOKEN or TOKEN_EXPIRED.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: INSUFFICIENT_SCOPE.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Meeting not found for this client.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: State, room slot, capacity, or creation-in-progress conflict.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Provider contract or validation rejection.
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Provider authentication, authorization, endpoint, response,
            or unknown-state failure.
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Provider unavailable or rate limited.
...
components:
  schemas:
...
    Class:
      type: object
      properties:
        id:
          type: string
          maxLength: 160
        date:
          type: string
          format: date
      required:
      - date
      - id
    ClassInfo:
      type: object
      properties:
        classId:
          type: string
        teacher:
          $ref: '#/components/schemas/Reference'
        batch:
          $ref: '#/components/schemas/Reference'
      required:
      - batch
      - classId
      - teacher
...
    CreateMeeting:
      type: object
      properties:
        meetingTitle:
          type: string
          maxLength: 250
        teacher:
          $ref: '#/components/schemas/Reference'
        batch:
          $ref: '#/components/schemas/Reference'
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        class:
          $ref: '#/components/schemas/Class'
      required:
      - batch
      - class
      - endAt
      - meetingTitle
      - startAt
      - teacher
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
        details: {}
        requestId:
          type: string
      required:
      - code
      - details
      - message
      - requestId
      - success
...
    MeetingData:
      type: object
      properties:
        id:
          type: string
          format: uuid
        classInfo:
          $ref: '#/components/schemas/ClassInfo'
        meetingInfo:
          $ref: '#/components/schemas/MeetingInfo'
        convay:
          $ref: '#/components/schemas/ConvayInfo'
        createdAt:
          type: string
          format: date-time
      required:
      - classInfo
      - convay
      - createdAt
      - id
      - meetingInfo
    MeetingInfo:
      type: object
      properties:
        meetingTitle:
          type: string
        classDate:
          type: string
          format: date
        scheduleType:
          $ref: '#/components/schemas/ScheduleTypeEnum'
        startAt:
          type: string
          format: date-time
          nullable: true
        endAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
      required:
      - classDate
      - endAt
      - meetingTitle
      - scheduleType
      - startAt
      - status
    MeetingResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/MeetingData'
      required:
      - data
      - success
...
    Reference:
      type: object
      properties:
        id:
          type: string
          maxLength: 160
        name:
          type: string
          maxLength: 200
      required:
      - id
      - name
...
    SearchData:
      type: object
      properties:
        count:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/MeetingData'
      required:
      - count
      - page
      - pageSize
      - results
    SearchResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/SearchData'
      required:
      - data
      - success
...
  securitySchemes:
    GatewayBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Obtain a Gateway accessToken from POST /api/v1/auth/token/. Default
        lifetime: 60 minutes (3600 seconds). Obtain another using client credentials
        near expiry or after 401 TOKEN_EXPIRED, then retry. Paste the Gateway token
        here. Convay tokens have a separate provider-controlled lifetime.'